requirements.txt
gradio
pillow
safetensors
timm
torch
torchvision